+++ vector_quantization/codebooks.py
@@ -1,4 +1,4 @@
-from dataclasses import asdict, dataclass, field, replace
+from dataclasses import asdict, dataclass, field, is_dataclass
 from functools import partial
 
 import torch
@@ -71,6 +71,11 @@ class CodebookParams:
     learnable_codebook: bool = False
     gumbel_params: GumbelParams = field(default_factory=GumbelParams)
     ema_update: bool = True
+    use_affine: bool = False
+    affine_params: AffineParameters = None
+    transform_input: str = "identity"
+    use_cosine_sim: bool = False
+    weights_regularization: str = "identity"
 
 
 @dataclass
@@ -99,7 +104,7 @@ class CosineSimCodebookParams:
     ema_update: bool = True
 
 
-class EuclideanCodebook(Module):
+class Codebook(Module):
     def __init__(
         self,
         dim,
@@ -118,20 +123,53 @@ class EuclideanCodebook(Module):
         ema_update: bool = True,
         use_affine: bool = False,
         affine_params: AffineParameters = None,
+        transform_input: str = "identity",
+        use_cosine_sim: bool = False,
+        weights_regularization: str = "identity",
     ):
         super().__init__()
-        self.transform_input = identity
+        if transform_input == "identity":
+            self.transform_input = identity
+        elif transform_input == "l2norm":
+            self.transform_input = l2norm
+        else:
+            raise f"The option {transform_input} as transform input function is not yet implemented"
+
+        if weights_regularization == "identity":
+            self.weights_regularization = identity
+        elif weights_regularization == "l2norm":
+            self.weights_regularization = l2norm
+        else:
+            raise f"The option {weights_regularization} for weights regularization is not yet implemented."
+
+        self.use_cosine_sim = use_cosine_sim
+        if use_cosine_sim:
+
+            def similarity_fn(x, y):
+                return einsum("h n d, h c d -> h n c", x, y)
+
+            self.similarity_fn = similarity_fn
+        else:
+
+            def similarity_fn(x, y):
+                return -cdist(x, y)
+
+            self.similarity_fn = similarity_fn
 
         self.decay = decay
         self.ema_update = ema_update
 
         init_fn = uniform_init if not initialization_by_kmeans else torch.zeros
-        embeddings = init_fn(num_codebooks, codebook_size, dim)
+        embeddings = self.weights_regularization(
+            init_fn(num_codebooks, codebook_size, dim)
+        )
 
         self.codebook_size = codebook_size
         self.num_codebooks = num_codebooks
 
-        self.kmeans_params = kmeans_params
+        self.kmeans_params = (
+            asdict(kmeans_params) if is_dataclass(kmeans_params) else kmeans_params
+        )
         self.eps_for_smoothing = eps_for_smoothing
         self.threshold_ema_dead_code = threshold_ema_dead_code
         self.reset_cluster_size = (
@@ -139,11 +177,11 @@ class EuclideanCodebook(Module):
             if reset_cluster_size is not None
             else threshold_ema_dead_code
         )
-
-        self.sample_fn_training = partial(gumbel_sample, **asdict(gumbel_params))
-        self.sample_fn_val = partial(
-            gumbel_sample, **asdict(replace(gumbel_params, training=False))
-        )
+        if is_dataclass(gumbel_params):
+            gumbel_params = asdict(gumbel_params)
+        self.sample_fn_training = partial(gumbel_sample, **gumbel_params)
+        gumbel_params["training"] = False
+        self.sample_fn_val = partial(gumbel_sample, **gumbel_params)
 
         assert not (
             use_ddp and num_codebooks > 1 and initialization_by_kmeans
@@ -151,19 +189,19 @@ class EuclideanCodebook(Module):
 
         self.sample_fn = (
             sample_vectors_distributed
-            if use_ddp and self.kmeans_params.sync
+            if use_ddp and self.kmeans_params["sync"]
             else batched_sample_vectors
         )
 
         self.distributed_replace_codes = distributed_replace_codes
         self.replace_sample_fn = (
             sample_vectors_distributed
-            if use_ddp and self.kmeans_params.sync and distributed_replace_codes
+            if use_ddp and self.kmeans_params["sync"] and distributed_replace_codes
             else batched_sample_vectors
         )
 
         self.kmeans_all_reduce_fn = (
-            distributed.all_reduce if use_ddp and self.kmeans_params.sync else noop
+            distributed.all_reduce if use_ddp and self.kmeans_params["sync"] else noop
         )
         self.all_reduce_fn = distributed.all_reduce if use_ddp else noop
 
@@ -180,15 +218,18 @@ class EuclideanCodebook(Module):
         # affine related params
 
         self.use_affine = use_affine
-        self.affine_params = affine_params
+        if use_affine:
+            self.affine_params = affine_params
 
-        self.register_buffer("batch_mean", None)
-        self.register_buffer("batch_variance", None)
+            self.register_buffer("batch_mean", None)
+            self.register_buffer("batch_variance", None)
 
-        self.register_buffer("codebook_mean_needs_init", torch.Tensor([True]))
-        self.register_buffer("codebook_mean", torch.empty(num_codebooks, 1, dim))
-        self.register_buffer("codebook_variance_needs_init", torch.Tensor([True]))
-        self.register_buffer("codebook_variance", torch.empty(num_codebooks, 1, dim))
+            self.register_buffer("codebook_mean_needs_init", torch.Tensor([True]))
+            self.register_buffer("codebook_mean", torch.empty(num_codebooks, 1, dim))
+            self.register_buffer("codebook_variance_needs_init", torch.Tensor([True]))
+            self.register_buffer(
+                "codebook_variance", torch.empty(num_codebooks, 1, dim)
+            )
 
     @torch.jit.ignore
     def initialize_embeddings(self, data, mask=None):
@@ -196,10 +237,12 @@ class EuclideanCodebook(Module):
             c = data.shape[0]
             data = rearrange(data[mask], "(c n) d -> c n d", c=c)
 
+        # note that if use_cosine_sim is True, then the embeddings are already "regularized"
         embeddings, cluster_size = kmeans(
             data,
             num_clusters=self.codebook_size,
-            num_iters=self.kmeans_params.iter,
+            num_iters=self.kmeans_params["iter"],
+            use_cosine_sim=self.use_cosine_sim,
             sample_fn=self.sample_fn,
             all_reduce_fn=self.kmeans_all_reduce_fn,
         )
@@ -209,6 +252,33 @@ class EuclideanCodebook(Module):
         self.embeddings.data.copy_(embeddings)
         self.embed_avg.data.copy_(embed_sum)
         self.cluster_size.data.copy_(cluster_size)
+
+    def replace_codes(self, batch_samples, batch_mask):
+        batch_samples = self.weights_regularization(batch_samples)
+        for ind, (samples, mask) in enumerate(zip(batch_samples, batch_mask)):
+            sampled = self.replace_sample_fn(
+                rearrange(samples, "... -> 1 ..."), mask.sum().item()
+            )
+            sampled = rearrange(sampled, "1 ... -> ...")
+
+            if not self.distributed_replace_codes:
+                sampled = maybe_distributed_mean(sampled)
+
+            self.embeddings.data[ind][mask] = sampled
+            self.cluster_size.data[ind][mask] = self.reset_cluster_size
+            self.embed_avg.data[ind][mask] = sampled * self.reset_cluster_size
+
+    def expire_codes_(self, batch_samples):
+        if self.threshold_ema_dead_code == 0:
+            return
+
+        expired_codes = self.cluster_size < self.threshold_ema_dead_code
+
+        if not torch.any(expired_codes):
+            return
+
+        batch_samples = rearrange(batch_samples, "h ... d -> h (...) d")
+        self.replace_codes(batch_samples, batch_mask=expired_codes)
 
     @torch.jit.ignore
     def update_with_decay(self, buffer_name, new_value, decay):
@@ -303,32 +373,6 @@ class EuclideanCodebook(Module):
             self.affine_params.batch_decay,
         )
 
-    def replace(self, batch_samples, batch_mask):
-        for ind, (samples, mask) in enumerate(zip(batch_samples, batch_mask)):
-            sampled = self.replace_sample_fn(
-                rearrange(samples, "... -> 1 ..."), mask.sum().item()
-            )
-            sampled = rearrange(sampled, "1 ... -> ...")
-
-            if not self.distributed_replace_codes:
-                sampled = maybe_distributed_mean(sampled)
-
-            self.embeddings.data[ind][mask] = sampled
-            self.cluster_size.data[ind][mask] = self.reset_cluster_size
-            self.embed_avg.data[ind][mask] = sampled * self.reset_cluster_size
-
-    def expire_codes_(self, batch_samples):
-        if self.threshold_ema_dead_code == 0:
-            return
-
-        expired_codes = self.cluster_size < self.threshold_ema_dead_code
-
-        if not torch.any(expired_codes):
-            return
-
-        batch_samples = rearrange(batch_samples, "h ... d -> h (...) d")
-        self.replace(batch_samples, batch_mask=expired_codes)
-
     @autocast(enabled=False)
     def forward(self, x, mask=None, freeze_codebook=False):
         needs_codebook_dim = x.ndim < 4
@@ -365,12 +409,11 @@ class EuclideanCodebook(Module):
                 batch_std / codebook_std
             ) + self.batch_mean
 
-        dist = -cdist(flatten, embeddings)
+        similarities = self.similarity_fn(flatten, embeddings)
 
         embed_ind, embed_onehot = self.sample_fn_training(
-            dist,
+            similarities,
         )
-
         embed_ind = unpack_one(embed_ind, ps, "h *")
 
         if self.training:
@@ -404,6 +447,7 @@ class EuclideanCodebook(Module):
             ) * self.cluster_size.sum(dim=-1, keepdim=True)
 
             embed_normalized = self.embed_avg / rearrange(cluster_size, "... -> ... 1")
+            embed_normalized = self.weights_regularization(embed_normalized)
             self.embeddings.data.copy_(embed_normalized)
             self.expire_codes_(x)
 
@@ -412,21 +456,61 @@ class EuclideanCodebook(Module):
                 lambda t: rearrange(t, "1 ... -> ..."), (quantize, embed_ind)
             )
 
-        dist = unpack_one(dist, ps, "h * d")
+        similarities = unpack_one(similarities, ps, "h * d")
 
-        return quantize, embed_ind, dist
+        return quantize, embed_ind, similarities
 
 
-class CosineSimCodebook(Module):
+class EuclideanCodebook(Codebook):
     def __init__(
         self,
-        dim: int,
-        codebook_size: int,
-        num_codebooks: int = 1,
+        dim,
+        codebook_size,
+        num_codebooks=1,
         initialization_by_kmeans: bool = False,
         kmeans_params: KmeansParameters = None,
         decay: float = 0.8,
-        eps_for_smoothing: float = 1e-5,
+        eps_for_smoothing: float = 0.00001,
+        threshold_ema_dead_code: int = 2,
+        reset_cluster_size: int = None,
+        use_ddp: bool = False,
+        distributed_replace_codes: bool = True,
+        learnable_codebook: bool = False,
+        gumbel_params: GumbelParams = GumbelParams(),
+        ema_update: bool = True,
+        use_affine: bool = False,
+        affine_params: AffineParameters = None,
+    ):
+        super().__init__(
+            dim,
+            codebook_size,
+            num_codebooks,
+            initialization_by_kmeans,
+            kmeans_params,
+            decay,
+            eps_for_smoothing,
+            threshold_ema_dead_code,
+            reset_cluster_size,
+            use_ddp,
+            distributed_replace_codes,
+            learnable_codebook,
+            gumbel_params,
+            ema_update,
+            use_affine,
+            affine_params,
+        )
+
+
+class CosineSimCodebook(Codebook):
+    def __init__(
+        self,
+        dim,
+        codebook_size,
+        num_codebooks=1,
+        initialization_by_kmeans: bool = False,
+        kmeans_params: KmeansParameters = None,
+        decay: float = 0.8,
+        eps_for_smoothing: float = 0.00001,
         threshold_ema_dead_code: int = 2,
         reset_cluster_size: int = None,
         use_ddp: bool = False,
@@ -435,180 +519,22 @@ class CosineSimCodebook(Module):
         gumbel_params: GumbelParams = GumbelParams(),
         ema_update: bool = True,
     ):
-        super().__init__()
-        self.transform_input = l2norm
-
-        self.ema_update = ema_update
-        self.decay = decay
-
-        if not initialization_by_kmeans:
-            embeddings = l2norm(uniform_init(num_codebooks, codebook_size, dim))
-        else:
-            embeddings = torch.zeros(num_codebooks, codebook_size, dim)
-
-        self.codebook_size = codebook_size
-        self.num_codebooks = num_codebooks
-
-        self.kmeans_params = kmeans_params
-        self.eps_for_smoothing = eps_for_smoothing
-        self.threshold_ema_dead_code = threshold_ema_dead_code
-        self.reset_cluster_size = (
-            reset_cluster_size
-            if reset_cluster_size is not None
-            else threshold_ema_dead_code
-        )
-
-        self.sample_fn_training = partial(gumbel_sample, **asdict(gumbel_params))
-        gumbel_params_val = gumbel_params
-        gumbel_params_val.trinaing = False
-        self.sample_fn_val = partial(gumbel_sample, **asdict(gumbel_params_val))
-
-        self.sample_fn = (
-            sample_vectors_distributed
-            if use_ddp and self.kmeans_params.sync
-            else batched_sample_vectors
-        )
-
-        self.distributed_replace_codes = distributed_replace_codes
-        self.replace_sample_fn = (
-            sample_vectors_distributed
-            if use_ddp and self.kmeans_params.sync and distributed_replace_codes
-            else batched_sample_vectors
-        )
-
-        self.kmeans_all_reduce_fn = (
-            distributed.all_reduce if use_ddp and self.kmeans_params.sync else noop
-        )
-        self.all_reduce_fn = distributed.all_reduce if use_ddp else noop
-
-        self.is_initialized = not initialization_by_kmeans
-        self.register_buffer("cluster_size", torch.zeros(num_codebooks, codebook_size))
-        self.register_buffer("embed_avg", embeddings.clone())
-
-        self.learnable_codebook = learnable_codebook
-        if learnable_codebook:
-            self.embeddings = nn.Parameter(embeddings)
-        else:
-            self.register_buffer("embeddings", embeddings)
-
-    @torch.jit.ignore
-    def initialize_embeddings(self, data, mask=None):
-        if mask is not None:
-            c = data.shape[0]
-            data = rearrange(data[mask], "(c n) d -> c n d", c=c)
-
-        embeddings, cluster_size = kmeans(
-            data,
-            self.codebook_size,
-            self.kmeans_params.iter,
+        super().__init__(
+            dim,
+            codebook_size,
+            num_codebooks,
+            initialization_by_kmeans,
+            kmeans_params,
+            decay,
+            eps_for_smoothing,
+            threshold_ema_dead_code,
+            reset_cluster_size,
+            use_ddp,
+            distributed_replace_codes,
+            learnable_codebook,
+            gumbel_params,
+            ema_update,
+            transform_input="l2norm",
             use_cosine_sim=True,
-            sample_fn=self.sample_fn,
-            all_reduce_fn=self.kmeans_all_reduce_fn,
+            weights_regularization="l2norm",
         )
-
-        embed_sum = embeddings * rearrange(cluster_size, "... -> ... 1")
-
-        self.embeddings.data.copy_(embeddings)
-        self.embed_avg.data.copy_(embed_sum)
-        self.cluster_size.data.copy_(cluster_size)
-
-    def replace(self, batch_samples, batch_mask):
-        batch_samples = l2norm(batch_samples)
-
-        for ind, (samples, mask) in enumerate(zip(batch_samples, batch_mask)):
-            sampled = self.replace_sample_fn(
-                rearrange(samples, "... -> 1 ..."), mask.sum().item()
-            )
-            sampled = rearrange(sampled, "1 ... -> ...")
-
-            if not self.distributed_replace_codes:
-                sampled = maybe_distributed_mean(sampled)
-
-            self.embeddings.data[ind][mask] = sampled
-            self.embed_avg.data[ind][mask] = sampled * self.reset_cluster_size
-            self.cluster_size.data[ind][mask] = self.reset_cluster_size
-
-    def expire_codes_(self, batch_samples):
-        if self.threshold_ema_dead_code == 0:
-            return
-
-        expired_codes = self.cluster_size < self.threshold_ema_dead_code
-
-        if not torch.any(expired_codes):
-            return
-
-        batch_samples = rearrange(batch_samples, "h ... d -> h (...) d")
-        self.replace(batch_samples, batch_mask=expired_codes)
-
-    @autocast(enabled=False)
-    def forward(self, x, mask=None, freeze_codebook=False):
-        needs_codebook_dim = x.ndim < 4
-
-        x = x.float()
-
-        if needs_codebook_dim:
-            x = rearrange(x, "... -> 1 ...")
-
-        flatten, ps = pack_one(x, "h * d")
-
-        if mask is not None:
-            mask = repeat(
-                mask,
-                "b n -> c (b h n)",
-                c=flatten.shape[0],
-                h=flatten.shape[-2] // (mask.shape[0] * mask.shape[1]),
-            )
-
-        if not self.is_initialized:
-            self.initialize_embeddings(flatten, mask=mask)
-            self.is_initialized = True
-
-        embeddings = (
-            self.embeddings if self.learnable_codebook else self.embeddings.detach()
-        )
-
-        dist = einsum("h n d, h c d -> h n c", flatten, embeddings)
-
-        embed_ind, embed_onehot = self.sample_fn_training(
-            dist,
-        )
-        embed_ind = unpack_one(embed_ind, ps, "h *")
-
-        if self.training:
-            unpacked_onehot = unpack_one(embed_onehot, ps, "h * c")
-            quantize = einsum("h b n c, h c d -> h b n d", unpacked_onehot, embeddings)
-        else:
-            quantize = batched_embedding(embed_ind, embeddings)
-
-        if self.training and self.ema_update and not freeze_codebook:
-            if mask is not None:
-                embed_onehot[~mask] = 0.0
-
-            bins = embed_onehot.sum(dim=1)
-            self.all_reduce_fn(bins)
-
-            ema_inplace(self.cluster_size.data, bins, self.decay)
-
-            embed_sum = einsum("h n d, h n c -> h c d", flatten, embed_onehot)
-            embed_sum = embed_sum.contiguous()
-            self.all_reduce_fn(embed_sum)
-
-            ema_inplace(self.embed_avg.data, embed_sum, self.decay)
-
-            cluster_size = laplace_smoothing(
-                self.cluster_size, self.codebook_size, self.eps_for_smoothing
-            ) * self.cluster_size.sum(dim=-1, keepdim=True)
-
-            embed_normalized = self.embed_avg / rearrange(cluster_size, "... -> ... 1")
-            embed_normalized = l2norm(embed_normalized)
-
-            self.embeddings.data.copy_(l2norm(embed_normalized))
-            self.expire_codes_(x)
-
-        if needs_codebook_dim:
-            quantize, embed_ind = map(
-                lambda t: rearrange(t, "1 ... -> ..."), (quantize, embed_ind)
-            )
-
-        dist = unpack_one(dist, ps, "h * d")
-        return quantize, embed_ind, dist